Split dashboard cards into smart cards and CardView components #396
Each dashboard card now delegates rendering to a pure presentational
*CardView component that receives prepared data through [Parameter]
values and injects no HTTP clients or backend-facing services. Smart
cards keep their existing self-loading behavior, so standalone usage
is unchanged. The three time-series cards already rendered through the
shared TimeSeriesValueCard view and needed no change.
