+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/Assets/AssetsDistributionOverviewCard.razor.cs
@@ -1,11 +1,8 @@
-using ApexCharts;
 using FinanceManager.Application.Identity.Users;
-using FinanceManager.Components.Helpers;
 using FinanceManager.Components.Models;
 using FinanceManager.Components.Services;
 using FinanceManager.Domain.Entities.Currencies;
 using FinanceManager.Domain.Entities.MoneyFlowModels;
-using FinanceManager.Domain.Providers;
 using FinanceManager.Domain.Services;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
@@ -14,16 +11,10 @@ namespace FinanceManager.Components.Components.Features.Dashboard.Cards.Assets;
 
 public partial class AssetsDistributionOverviewCard
 {
-    private const string _viewByType = "type";
-    private const string _viewByWallet = "wallet";
-
-    private string _view = _viewByType;
     private bool _isLoading;
     private Currency _currency = DefaultCurrency.PLN;
-    private decimal _totalAssets;
     private List<NameValueResult> _typeData = [];
     private List<NameValueResult> _walletData = [];
-    private ApexChart<NameValueResult>? _chart;
 
     [Parameter] public string Height { get; set; } = "300px";
     [Parameter] public DateTime StartDateTime { get; set; }
@@ -34,29 +25,9 @@ public partial class AssetsDistributionOverviewCard
     [Inject] public required ISettingsService SettingsService { get; set; }
     [Inject] public required ILoginService LoginService { get; set; }
 
-    private List<NameValueResult> ActiveData => _view == _viewByWallet ? _walletData : _typeData;
-
-    private readonly ApexChartOptions<NameValueResult> _chartOptions = new()
-    {
-        Chart = new Chart
-        {
-            Toolbar = new Toolbar { Show = false },
-            Background = "transparent",
-        },
-        Legend = new Legend { Show = false },
-        Colors = ColorsProvider.GetColors(),
-    };
-
     protected override void OnInitialized()
     {
         _currency = SettingsService.GetCurrency();
-        _chartOptions.Tooltip = new Tooltip
-        {
-            Y = new TooltipY
-            {
-                Formatter = ChartHelper.GetCurrencyFormatter(_currency.ShortName),
-            },
-        };
     }
 
     protected override async Task OnParametersSetAsync()
@@ -71,7 +42,6 @@ public partial class AssetsDistributionOverviewCard
             {
                 _typeData = [];
                 _walletData = [];
-                _totalAssets = 0;
                 return;
             }
 
@@ -86,10 +56,6 @@ public partial class AssetsDistributionOverviewCard
             var snapshot = await AssetsPageCardsCacheService.GetSnapshotAsync(context);
             _typeData = [.. snapshot.EndAssetsPerType];
             _walletData = [.. snapshot.EndAssetsPerAccount];
-            _totalAssets = _typeData.Count == 0 ? 0 : Math.Round(_typeData.Sum(x => x.Value), 2);
-
-            if (_chart is not null)
-                await _chart.UpdateSeriesAsync(true);
         }
         catch (Exception ex)
         {
@@ -100,13 +66,5 @@ public partial class AssetsDistributionOverviewCard
             _isLoading = false;
             StateHasChanged();
         }
-    }
-
-    private async Task OnViewChangedAsync(string view)
-    {
-        _view = view;
-        StateHasChanged();
-        if (_chart is not null)
-            await _chart.UpdateSeriesAsync(true);
     }
 }
