+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/ExpenseDistributionOverviewCard.razor.cs
@@ -1,10 +1,7 @@
-using ApexCharts;
 using FinanceManager.Application.Identity.Users;
-using FinanceManager.Components.Helpers;
 using FinanceManager.Components.HttpClients;
 using FinanceManager.Domain.Entities.Currencies;
 using FinanceManager.Domain.Entities.MoneyFlowModels;
-using FinanceManager.Domain.Providers;
 using FinanceManager.Domain.Services;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
@@ -16,9 +13,7 @@ public partial class ExpenseDistributionOverviewCard
 {
     private bool _isLoading;
     private Currency _currency = DefaultCurrency.PLN;
-    private decimal _totalExpenses;
     private List<NameValueResult> _data = [];
-    private ApexChart<NameValueResult>? _chart;
 
     [Parameter] public string Height { get; set; } = "300px";
     [Parameter] public DateTime StartDateTime { get; set; }
@@ -30,27 +25,9 @@ public partial class ExpenseDistributionOverviewCard
     [Inject] public required ISettingsService SettingsService { get; set; }
     [Inject] public required ILoginService LoginService { get; set; }
 
-    private readonly ApexChartOptions<NameValueResult> _chartOptions = new()
-    {
-        Chart = new Chart
-        {
-            Toolbar = new Toolbar { Show = false },
-            Background = "transparent",
-        },
-        Legend = new Legend { Show = false },
-        Colors = ColorsProvider.GetColors(),
-    };
-
     protected override void OnInitialized()
     {
         _currency = SettingsService.GetCurrency();
-        _chartOptions.Tooltip = new Tooltip
-        {
-            Y = new TooltipY
-            {
-                Formatter = ChartHelper.GetCurrencyFormatter(_currency.ShortName),
-            },
-        };
     }
 
     protected override async Task OnParametersSetAsync()
@@ -64,16 +41,11 @@ public partial class ExpenseDistributionOverviewCard
             if (user is null)
             {
                 _data = [];
-                _totalExpenses = 0;
                 return;
             }
 
             var data = await MoneyFlowHttpClient.GetExpenseDistribution(user.UserId, _currency, StartDateTime, EndDateTime);
             _data = [.. data];
-            _totalExpenses = _data.Count == 0 ? 0 : Math.Round(_data.Sum(x => x.Value), 2);
-
-            if (_chart is not null)
-                await _chart.UpdateSeriesAsync(true);
         }
         catch (Exception ex)
         {
