+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/ExpenseDistributionOverviewCardView.razor
@@ -1,0 +1,73 @@
+@using ApexCharts
+@using FinanceManager.Components.Components.Shared
+@using FinanceManager.Domain.Entities.MoneyFlowModels
+@using FinanceManager.Domain.Providers
+
+<MudCard Outlined Class="d-flex flex-column" Style="@($"height:{Height};")">
+    <MudPaper Class="d-flex align-start justify-space-between gap-4 pa-3 pb-0 flex-shrink-0" Elevation="0">
+        <div>
+            <MudText Typo="Typo.h5">Expense distribution</MudText>
+            @if (IsLoading)
+            {
+                <MudText Typo="Typo.h6" Class="text-muted">
+                    <StringSpinner CharactersCount="4" />
+                </MudText>
+            }
+            else if (TotalExpenses != 0)
+            {
+                <MudText Class="ed-total">@TotalExpenses.ToString("0.00") @CurrencyShortName</MudText>
+            }
+        </div>
+    </MudPaper>
+
+    <MudPaper Class="flex-grow-1 overflow-hidden d-flex align-center" Elevation="0">
+        @if (IsLoading)
+        {
+            <div style="width:100%;height:100%">
+                <ChartSpinner MaxHeight=@Height />
+            </div>
+        }
+        else if (Data.Any())
+        {
+            <div class="ed-body">
+                <div class="ed-chart-wrap">
+                    <ApexChart @ref="_chart" TItem="NameValueResult" Options="_chartOptions" Width="150" Height="150">
+                        <ApexPointSeries TItem="NameValueResult"
+                                         Items="Data"
+                                         Name="Expenses"
+                                         SeriesType="SeriesType.Pie"
+                                         XValue="e => e.Name"
+                                         YAggregate="@(e => e.Sum(x => Math.Round(x.Value, 2)))"
+                                         OrderByDescending="@(e => e.Y)" />
+                    </ApexChart>
+                </div>
+                <div class="ed-legend">
+                    @{
+                        var sorted = Data.OrderByDescending(x => x.Value).ToList();
+                        var colors = ColorsProvider.GetColors();
+                        for (int i = 0; i < sorted.Count; i++)
+                        {
+                            var item = sorted[i];
+                            var color = colors[i % colors.Count];
+                            var pct = TotalExpenses == 0 ? 0m : Math.Round(item.Value / TotalExpenses * 100m, 1);
+                            <div class="ed-legend-row">
+                                <span class="ed-legend-dot" style="background: @color"></span>
+                                <span class="ed-legend-name">@item.Name</span>
+                                <span class="ed-legend-vals">
+                                    <span class="ed-legend-amt">@item.Value.ToString("0.00")</span>
+                                    <span class="ed-legend-pct">@pct.ToString("0.0")%</span>
+                                </span>
+                            </div>
+                        }
+                    }
+                </div>
+            </div>
+        }
+        else
+        {
+            <div class="d-flex flex-grow-1 justify-center align-content-center flex-wrap">
+                <MudText Typo="Typo.button" HtmlTag="h3">No data</MudText>
+            </div>
+        }
+    </MudPaper>
+</MudCard>
