+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/FinancialInsightsCarousel.razor
@@ -1,103 +1,3 @@
-@using FinanceManager.Domain.Entities.Users
-
-<MudCard Outlined Class="d-flex flex-column overflow-hidden" Style="@($"height:{Height}")">
-    <div class="ins-inner"
-         @onmouseenter="OnMouseEnter" @onmouseleave="OnMouseLeave"
-         @onfocusin="OnFocusIn" @onfocusout="OnFocusOut">
-
-        @* ---- Header ---- *@
-        <div class="ins-head">
-            <span class="ins-eyebrow">Insights</span>
-            <span class="ins-spacer"></span>
-            @if (!_isLoading && CurrentInsight is not null)
-            {
-                <span class="ins-time">@GetRelativeTime(CurrentInsight.CreatedAt)</span>
-            }
-            <span class="ins-ai-pill">
-                <MudIcon Icon="@Icons.Material.Filled.AutoAwesome" Style="font-size:12px;line-height:1" />
-                AI
-            </span>
-        </div>
-
-        @* ---- Story progress bar (carousel mode only) ---- *@
-        @if (!_isLoading && _insights.Count > 1)
-        {
-            <div class="ins-progress @(_isPaused ? "paused" : "")" style="--dwell:6000ms">
-                @for (var k = 0; k < _insights.Count; k++)
-                {
-                    var kk = k;
-                    <span class="seg @(kk < _currentIndex ? "done" : kk == _currentIndex ? "active" : "")">
-                        <i @key="@(kk == _currentIndex ? $"a{_currentIndex}" : $"s{kk}")"></i>
-                    </span>
-                }
-            </div>
-        }
-
-        @* ---- Body ---- *@
-        <div class="ins-body">
-            @if (_isLoading)
-            {
-                <div class="ins-skel" style="height:18px;width:82%"></div>
-                <div class="ins-skel" style="height:11px;width:100%;margin-top:16px"></div>
-                <div class="ins-skel" style="height:11px;width:94%;margin-top:8px"></div>
-                <div class="ins-skel" style="height:11px;width:60%;margin-top:8px"></div>
-                <div class="ins-skel-spacer"></div>
-                <div class="ins-skel-chips">
-                    <div class="ins-skel" style="height:24px;width:62px;border-radius:9999px"></div>
-                    <div class="ins-skel" style="height:24px;width:52px;border-radius:9999px"></div>
-                </div>
-            }
-            else if (!_insights.Any())
-            {
-                <div class="ins-empty">
-                    <span class="ins-empty-orb">
-                        <MudIcon Icon="@Icons.Material.Filled.AutoAwesome" Style="font-size:26px" />
-                    </span>
-                    <span class="ins-empty-title">No insights yet</span>
-                    <span class="ins-empty-sub">
-                        As your accounts update, we'll surface short insights about your finances here.
-                    </span>
-                </div>
-            }
-            else
-            {
-                <div class="ins-fade" @key="_currentIndex">
-                    <div class="ins-insight-title">@CurrentInsight!.Title</div>
-                    <div class="ins-insight-msg">@CurrentInsight!.Message</div>
-                </div>
-                <div class="ins-body-spacer"></div>
-                @if (_insights.Count <= 1)
-                {
-                    <div class="ins-tags ins-tags-body">
-                        @foreach (var tag in GetTags(CurrentInsight!.Tags))
-                        {
-                            <span class="ins-tag">@tag</span>
-                        }
-                    </div>
-                }
-            }
-        </div>
-
-        @* ---- Footer: tags left + arrows right (carousel only) ---- *@
-        @if (!_isLoading && _insights.Count > 1)
-        {
-            <div class="ins-foot">
-                <div class="ins-tags">
-                    @foreach (var tag in GetTags(CurrentInsight!.Tags))
-                    {
-                        <span class="ins-tag">@tag</span>
-                    }
-                </div>
-                <div class="ins-nav">
-                    <button class="ins-arrow" aria-label="Previous insight" @onclick="GoPrevious">
-                        <MudIcon Icon="@Icons.Material.Filled.ChevronLeft" Style="font-size:20px" />
-                    </button>
-                    <button class="ins-arrow" aria-label="Next insight" @onclick="GoNext">
-                        <MudIcon Icon="@Icons.Material.Filled.ChevronRight" Style="font-size:20px" />
-                    </button>
-                </div>
-            </div>
-        }
-
-    </div>
-</MudCard>
+<FinancialInsightsCarouselView Height="@Height"
+                               IsLoading="_isLoading"
+                               Insights="_insights" />
