+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/FinancialInsightsCarouselView.razor.css
