+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/Liabilities/LiabilitiesDistributionOverviewCardView.razor.css
