+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/RecurringTransactionDetectorCard.razor
@@ -1,112 +1,9 @@
-@using FinanceManager.Domain.Entities.FinancialAccounts.Currencies
-
-<MudCard Outlined Class="d-flex flex-nowrap" Style="@($"height:{Height};")">
-    <MudPaper Class="flex-grow-0 px-4 pt-3 pb-1" Elevation="0">
-        @if (_selectedItem is null)
-        {
-            <MudText Typo=Typo.h5>Recurring transactions</MudText>
-            @if (_isLoading)
-            {
-                <MudText Typo=Typo.subtitle1 Class="text-muted">
-                    <StringSpinner CharactersCount="4" />
-                </MudText>
-            }
-            else if (_data.Count > 0)
-            {
-                <MudText Typo=Typo.subtitle1 Class="text-muted">@_totalMonthlySpend.ToString("0.00") PLN/mo</MudText>
-            }
-        }
-        else
-        {
-            <div class="d-flex align-center gap-1">
-                <MudIconButton Icon="@Icons.Material.Filled.ArrowBack" Size="Size.Small" OnClick="Back" aria-label="Back to recurring transactions" />
-                <MudText Typo=Typo.h5>@_selectedItem.Name</MudText>
-            </div>
-            @if (_detailEntries.Count > 0)
-            {
-                <MudText Typo=Typo.caption Class="text-muted">@_detailEntries.First().PostingDate.Year</MudText>
-            }
-        }
-    </MudPaper>
-
-    <MudPaper Class="flex-grow-1 overflow-auto" Elevation="0">
-        @if (_selectedItem is null)
-        {
-            @if (_isLoading)
-            {
-                <div style="height:100%">
-                    <ChartSpinner MaxHeight=@Height />
-                </div>
-            }
-            else if (_data.Count > 0)
-            {
-                <MudList T="string" Dense Class="px-2 py-0">
-                    @foreach (var item in _data)
-                    {
-                        <MudListItem Class="py-1 px-2 cursor-pointer" @onclick="() => SelectItem(item)">
-                            <div class="d-flex align-items-center gap-3 w-100">
-                                <MudIcon Icon="@Icons.Material.Filled.Repeat" Size="Size.Small" />
-                                <div class="flex-grow-1">
-                                    <MudText Typo="Typo.body2" Class="font-weight-500">@item.Name</MudText>
-                                    <MudProgressLinear Value="@GetPercentage(item.Value)" Color="Color.Warning"
-                                                      Class="mt-1" Size="Size.Small" />
-                                </div>
-                                <MudText Typo="Typo.body2" Color="Color.Warning" Class="font-weight-600"
-                                         Style="min-width: 120px; text-align: right;">
-                                    @($"{item.Value:0.00} PLN/mo")
-                                </MudText>
-                            </div>
-                        </MudListItem>
-                    }
-                </MudList>
-            }
-            else
-            {
-                <div class="d-flex flex-grow-1 justify-center align-content-center flex-wrap">
-                    <MudText Typo="Typo.button" HtmlTag="h3">No recurring transactions</MudText>
-                </div>
-            }
-        }
-        else
-        {
-            @if (_isLoadingDetail)
-            {
-                <div style="height:100%">
-                    <ChartSpinner MaxHeight=@Height />
-                </div>
-            }
-            else if (_detailEntries.Count > 0)
-            {
-                <MudTable Items="_detailEntries" Dense Hover Striped Elevation="0" Class="pa-2">
-                    <HeaderContent>
-                        <MudTh>Date</MudTh>
-                        <MudTh>Description</MudTh>
-                        <MudTh Style="text-align:right">Amount</MudTh>
-                    </HeaderContent>
-                    <RowTemplate>
-                        <MudTd DataLabel="Date">@context.PostingDate.ToString("MM-dd")</MudTd>
-                        <MudTd DataLabel="Description">
-                            @{
-                                var desc = string.IsNullOrEmpty(context.ContractorDetails) ? context.Description : context.ContractorDetails;
-                                var maxLength = 30;
-                                var displayDesc = desc.Length > maxLength ? desc.Substring(0, maxLength) + "..." : desc;
-                            }
-                            <MudTooltip Text="@desc" Arrow Placement="Placement.Bottom">
-                                <MudText>@displayDesc</MudText>
-                            </MudTooltip>
-                        </MudTd>
-                        <MudTd DataLabel="Amount" Style="text-align:right">
-                            <MudText Color="Color.Error">@($"{Math.Abs(context.ValueChange):0.00} PLN")</MudText>
-                        </MudTd>
-                    </RowTemplate>
-                </MudTable>
-            }
-            else
-            {
-                <div class="d-flex flex-grow-1 justify-center align-content-center flex-wrap">
-                    <MudText Typo="Typo.button" HtmlTag="h3">No entries found</MudText>
-                </div>
-            }
-        }
-    </MudPaper>
-</MudCard>
+<RecurringTransactionDetectorCardView Height="@Height"
+                                      IsLoading="_isLoading"
+                                      IsLoadingDetail="_isLoadingDetail"
+                                      Data="_data"
+                                      TotalMonthlySpend="_totalMonthlySpend"
+                                      SelectedItem="_selectedItem"
+                                      DetailEntries="_detailEntries"
+                                      OnItemSelected="SelectItem"
+                                      OnBack="Back" />
