+++ code/FinanceManager.Components/Components/Features/Dashboard/Cards/RecurringTransactionDetectorCard.razor.cs
@@ -101,11 +101,4 @@ public partial class RecurringTransactionDetectorCard
         _detailEntries = [];
         _isLoadingDetail = false;
     }
-
-    private double GetPercentage(decimal value)
-    {
-        if (_data.Count == 0) return 0;
-        var max = _data.Max(x => x.Value);
-        return max == 0 ? 0 : (double)(value / max * 100);
-    }
 }
